Initial work on the buffer storage example, may be abandoned
This example will work, but it feels like it emphasizes the binary order
format way too much.  If we were actually doing this much with the
orders in node, we'd probably parse the buffer out into a pile of UInts
and then drop the buffer, thus bypassing the problem I'm trying to demo.